Adding 'mousemove' event listener
Adding 'mousemove' event listener to be executed once.
Why? When I use a masony plugin gallery, it can take a while for the page to be fully loaded, andsome images in the viewport haven't loaded. By adding this event, it fire the 'check' function when the user move the cursor over the body.
I've wrapped the listener in a separeted function to be executed once.